refactor: separate files for diary dashboard page
