+++ src/static/diary/dashboard/styles.css
